[Added missing pickup scripts]
The PickupInventory was added into the scripts folder inside the project, it is located in the network/player/ folder.

The CollectablePickup was added into the scripts folder which enables pickups to be collectable if the player has a 'PickupInventory' component on themselves.

AbstractPickup and TimedPickup were modified in order to work with CollectablePickup.