[Fixed character selection over network, fixed Kris' UI over network, moved BlankPlayerSetup.cs to more appropriate location]
When character selection branch was integrated into master, some changes to code occurred which ended up breaking selection over the network.
Some of these changes included the removal of a [Command] function. This has been rectified and has been tested over a network.

Once this was fixed, it was noticed that the new UI Kris implemented wasn't client-side. All clients could see each other's UI. This has also been fixed by re-disabling the canvas in all player prefabs as it was before. The canvas is enabled client-side through code already to workaround this issue. Has been tested over a network and is fixed.