[Added Menus]
The game now uses a menu to navigate into the game. The original network manager HUD has been hidden, not removed just in case we need it later.
Every UI element has also been anchored so screen resolution shouldn't affect the UI now.
The UI will be improved upon as the game gets more developed and the options needed in the game are given more details.