[Player Kaiju now spawns directly, Blank Player removed]
After looking at the base code of the Network Manager, it was discovered how Unity spawns a player. Using that, it is now possible to spawn a kaiju directly, rather than having to use a 'Blank Player' as a stepping stone to spawning the desired kaiju.

_______________________________
Changes made in CustomNetworkManager.cs

When the server starts, it registers a function to be called when the server receives an ID number from the client. This function will receive a message containing the name of the kaiju selected plus networking info such as the client's ID etc.

Functions that are called automatically when a client connects to the server have been overrided for the purpose of implementing a custom player spawn system. The ID number is sent to the server in these functions.
When the function is called, it checks the name of the kaiju from the client's playerprefs and spawns the corresponding kaiju.

A (REQUIRED)  'CustomAddPlayerMessage' class which inherits from 'MessageBase' has been created for the registered function to run. This class holds the name of the kaiju.

_____________________________
Changes made in Offline Scene:

Custom Network Manager has auto-spawn disabled, the player prefab field cleared and kaijus added to new GameObject variables created.
SelectionCanvas' buttons have been edited to target a new function in Custom Network Manager which adds the selected kaiju's name to PlayerPrefs.