[When the game ends, the server now properly resets the OnlineScene and disconnects all clients.]
When a player reaches 200 points, all clients are told which player has won.
After a few seconds, the server kicks all clients back to the main menu.
The server then reloads the OnlineScene/resets the game.
When clients connect to the game, they are brought into the new OnlineScene.with no destroyed buildings etc.