[Fixed bug with server-reset feature where after the server has reset about 5 times it can't add any new players]
There was a function in the Custom Network Manager that handled random positions for a player to spawn at. This list was reduced each time a player spawned, and soon enough there were no spawn positions to choose from and the manager could not spawn a player. The line of code that removes elements from the random spawn list has been commented out.