[Added game timer to game. Added handling for a situation where a client joins the game but is not given a player]
When the game timer reaches 0, the person with the highest score is displayed. If there is a tie, then text saying there is a tie is displayed.
The clients then disconnect back to the main menu screen and the server reloads the level.

In a situation where a client connects to the server but the server cannot spawn a player, the server reloads the game scene and properly spawns a player for all clients.